@@ -2,10 +2,19 @@ require('dotenv').config();
 const { Sequelize } = require('sequelize');
 const fs = require('fs');
 const path = require('path');
-const { DB_USER, DB_PASSWORD, DB_HOST } = process.env;
+const { DB_USER, DB_PASSWORD, DB_HOST,DB_NAME,DB_DEPLOY } = process.env;
 
+/*
 const sequelize = new Sequelize(
-   `postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}/pokemon`,
+   `postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}/${DB_NAME}`,
+   {
+      logging: false, // set to console.log to see the raw SQL queries
+      native: false, // lets Sequelize know we can use pg-native for ~30% more speed
+   }
+);
+*/
+const sequelize = new Sequelize(
+   DB_DEPLOY,
    {
       logging: false, // set to console.log to see the raw SQL queries
       native: false, // lets Sequelize know we can use pg-native for ~30% more speed
@@ -15,8 +24,10 @@ const basename = path.basename(__filename);
 
 const modelDefiners = [];
 
-// Leemos todos los archivos de la carpeta Models, los requerimos y agregamos al arreglo modelDefiners
-fs.readdirSync(path.join(__dirname, '/models'))
+//manejo de errores de lectura de los archivos de modelos
+try{
+   // Leemos todos los archivos de la carpeta Models, los requerimos y agregamos al arreglo modelDefiners
+   fs.readdirSync(path.join(__dirname, '/models'))
    .filter(
       (file) =>
          file.indexOf('.') !== 0 &&
@@ -26,6 +37,11 @@ fs.readdirSync(path.join(__dirname, '/models'))
    .forEach((file) => {
       modelDefiners.push(require(path.join(__dirname, '/models', file)));
    });
+} catch(error){
+   console.error('Error en la lectura de los archivos de modelos:', err);
+}
+
+
 
 // Injectamos la conexion (sequelize) a todos los modelos
 modelDefiners.forEach((model) => model(sequelize));
@@ -39,10 +55,19 @@ sequelize.models = Object.fromEntries(capsEntries);
 
 // En sequelize.models están todos los modelos importados como propiedades
 // Para relacionarlos hacemos un destructuring
-const { Pokemon } = sequelize.models;
+const { Pokemon,Type } = sequelize.models;
 
 // Aca vendrian las relaciones
+Pokemon.belongsToMany(Type, { through: "PokemonType" });
+Type.belongsToMany(Pokemon, { through: "PokemonType" });
 // Product.hasMany(Reviews);
+
+//Cortar la conexion a la base de datos al cerrar la aplicacion
+process.on('exit', () => {
+   sequelize.close().then(() => {
+     console.log('Conexion a la base de datos cerrada');
+   });
+ });
 
 module.exports = {
    ...sequelize.models, // para poder importar los modelos así: const { Product, User } = require('./db.js');
