@@ -1,9 +1,31 @@
-import './App.css';
+import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom';
+import LandingPage from './views/LandingPage/LandingPage';
+import Home from './views/Home/Home';
+import Detail from './views/Detail/Detail';
+import Form from './views/Form/Form';
+import NavBar from './componentes/NavBar/NavBar';
 
 function App() {
   return (
-    <div className="App">
-      <h1>Henry Pokemon</h1>
+    <Router>
+      <AppContent />
+    </Router>
+  );
+}
+
+function AppContent() {
+  const location = useLocation();
+
+  return (
+    <div>
+      {location.pathname !== '/' && <NavBar />} 
+      <Routes>
+        
+        <Route path='/' element={<LandingPage />} />
+        <Route path='/home' element={<Home />} />
+        <Route path='/detail/:detailId' element={<Detail />} />
+        <Route path='/create' element={<Form />} /> 
+      </Routes>
     </div>
   );
 }
